% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclismProj-package.R
\docType{package}
\name{cyclismProj-package}
\alias{cyclismProj}
\alias{cyclismProj-package}
\title{Cyclists research project}
\description{
R-package containing the main elements related to the cyclist research
project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nim-ach/cyclismProj}
  \item Report bugs at \url{https://github.com/nim-ach/cyclismProj/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Katherine Harris King \email{katherine.harris@umag.cl}
  \item Valeska Oyarzo Igor \email{valeska.oyarzo@umag.cl}
  \item Cristian Núñez Espinosa \email{cristian.nunez@umag.cl} (\href{https://orcid.org/0000-0002-9896-7062}{ORCID})
}

}
\keyword{internal}
