% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cyclist}
\alias{cyclist}
\title{Data from the Cyclism study from CADI-UMAG}
\format{
A data frame with 42 rows y 46 variables:

\itemize{
\item{\code{id}: Factor. ID for each subject.}
\item{\code{age}: Numeric. Age in years.}
\item{\code{sex}: Factor. Participant's biological sex.}
\item{\code{weight}: Numeric. Weight in kilograms.}
\item{\code{height}: Numeric. Height in centimeters.}
\item{\code{bmi}: Numeric. Body mass index.}
\item{\code{imo}: Numeric. Bone muscle index.}
\item{\code{waist_hip_ratio}: Numeric. Waist to hip ratio.}
\item{\code{fat_mass_kg}: Numeric. Fat mass in kilograms.}
\item{\code{muscle_mass_kg}: Numeric. Muscle mass in kilograms.}
\item{\code{bone_mass_kg}: Numeric. Bone mass in kilograms.}
\item{\code{residual_mass_kg}: Numeric. Residual mass in kilograms.}
\item{\code{fat_mass_perc}: Numeric. Fat mass in percentage.}
\item{\code{muscle_mass_perc}: Numeric. Muscle mass in percentage.}
\item{\code{bone_mass_perc}: Numeric. Bone mass in percentage.}
\item{\code{residual_mass_perc}: Numeric. Residual mass in percentage.}
\item{\code{mme}: Numeric. Skeletal muscle mass.}
\item{\code{tmb}: Numeric. Basal metabolic rate.}
\item{\code{waist_girth}: Numeric. Waist circumference.}
\item{\code{visceral_fat}: Numeric. Visceral fat.}
\item{\code{bone_mineral_mass}: Numeric. Bone mineral mass in kilograms.}
\item{\code{celular_mass}: Numeric. Celular mass in kilograms.}
\item{\code{arm_girth}: Numeric. Arm circumference.}
\item{\code{act_mlg}: Numeric. description.}
\item{\code{ime}: Numeric. description.}
\item{\code{water_aec_act}: Numeric. description.}
\item{\code{water_aic}: Numeric. description.}
\item{\code{water_act}: Numeric. description.}
\item{\code{water_aec}: Numeric. description.}
\item{\code{spaq_ssi}: Numeric. description.}
\item{\code{spaq_severity}: Numeric. description.}
\item{\code{beck_dep_score}: Numeric. description.}
\item{\code{beck_dep_cat}: Factor. description.}
\item{\code{beck_anx_score}: Numeric. description.}
\item{\code{beck_anx_cat}: Factor. description.}
\item{\code{hrv_basal_pns}: Numeric. description.}
\item{\code{hrv_basal_stress}: Numeric. description.}
\item{\code{hrv_basal_sns}: Numeric. description.}
\item{\code{hrv_post_pns}: Numeric. description.}
\item{\code{hrv_post_stress}: Numeric. description.}
\item{\code{hrv_post_sns}: Numeric. description.}
\item{\code{ftp_mean_power}: Numeric. description.}
\item{\code{ftp_w_kg}: Numeric. description.}
\item{\code{hrv_delta_pns}: Numeric. description.}
\item{\code{hrv_delta_stress}: Numeric. description.}
\item{\code{hrv_delta_sns}: Numeric. description.}
}
}
\source{
CADI-UMAG, Punta Arenas, Chile.
}
\usage{
cyclist
}
\description{
This dataset contains anthropometric, performance and autonomic variables.
A subset of this dataset using only males subjects can be observed at \link[=males]{males}.
}
\seealso{
\itemize{
\item \link[=males]{males} dataset.
}
}
\keyword{datasets}
