% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{males}
\alias{males}
\title{Subset data from the Cyclism study from CADI-UMAG}
\format{
A data frame with 12 rows y 45 variables
}
\source{
CADI-UMAG, Punta Arenas, Chile.
}
\usage{
males
}
\description{
This is a subset of the dataset using only males subjects from the \link[=cyclist]{cyclist}
dataset.
}
\seealso{
\itemize{
\item \link[=cyclist]{cyclist} dataset.
}
}
\keyword{datasets}
