% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{mean_ic_boot}
\alias{mean_ic_boot}
\title{Media e intervalo de confianza usando bootstrap}
\usage{
mean_ic_boot(x, nboot = 5000, ci = 0.95, seed = 1234, k = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{nboot}{Number of bootstrap resamples.}

\item{ci}{Confidence interval.}

\item{seed}{Reproducibility seed.}

\item{k}{Number of digits when printing the results.}
}
\description{
Media e intervalo de confianza usando bootstrap
}
